+++ include/printer/dpnl_printer.h
@@ -1,0 +1,11 @@
+/* DPNL Printer Header */
+
+#ifndef DPNL_PRINTER_H
+#define DPNL_PRINTER_H
+
+typedef struct _dbg dbg_t;
+typedef struct _debugwindow debugwindow_t;
+
+extern void print_dpnl(debugwindow_t *window, dbg_t *dbg);
+
+#endif