+++ include/printer/gpnl_printer.h
@@ -1,0 +1,11 @@
+/* GPNL Printer Header */
+
+#ifndef GPNL_PRINTER_H
+#define GPNL_PRINTER_H
+
+typedef struct _dbg dbg_t;
+typedef struct _debugwindow debugwindow_t;
+
+extern void print_gpnl(debugwindow_t *window, dbg_t *dbg);
+
+#endif