+++ include/printer/propcol.h
@@ -1,0 +1,12 @@
+/* PropCol Header */
+
+#ifndef PROPCOL_H
+#define PROPCOL_H
+
+// Project Headers (enum needed)
+#include "cli/debug/color.h"
+#include "system/util/memprop.h"
+
+extern COLOR propcol(const MEMPROP prop);
+
+#endif