+++ include/printer/rpnl_printer.h
@@ -1,0 +1,11 @@
+/* RPNL Printer Header */
+
+#ifndef RPNL_PRINTER_H
+#define RPNL_PRINTER_H
+
+typedef struct _dbg dbg_t;
+typedef struct _debugwindow debugwindow_t;
+
+extern void print_rpnl(debugwindow_t *window, dbg_t *dbg);
+
+#endif