+++ include/system/util/memprop.h
@@ -1,0 +1,14 @@
+/* Memory Properties Header */
+
+#ifndef MEMPROP_H
+#define MEMPROP_H
+
+typedef enum {
+    
+    NONE,
+    SRC,
+    DEST
+    
+} MEMPROP;
+
+#endif